feat: Add handleTabCompletion function for tab completion in terminal
This commit adds the `handleTabCompletion` function to the `tabCompletion.ts` file. The function takes a command as input and updates the command state based on the available commands that start with the input. If there is only one matching command, it sets the command state to that command. This function will be used for tab completion in the terminal-themed website.